feat: add department, employee, user functional

+++ src/api/v1/services/position.py
@@ -1,6 +1,7 @@
 from utils.service import BaseService
 from utils.unit_of_work import transaction_mode
 
+from schemas.auth import UserTokenSchema
 from schemas.position import (
     PositionCreateSchema,
     PositionUpdateSchema,
@@ -23,10 +24,16 @@ logger = logging.getLogger(__name__)
 class PositionService(BaseService):
 
     @transaction_mode
-    async def create_position(self, schema: PositionCreateSchema) -> PositionResponseSchema:
+    async def create_position(self,
+                              schema: PositionCreateSchema,
+                              user_token_schema: UserTokenSchema) -> PositionResponseSchema:
         company = await self.uow.company.get_by_query_one_or_none(id=schema.company_id)
+
         if not company:
             raise BadRequestException(detail="Такой компании не существует")
+
+        if company.id != user_token_schema.company_id:
+            raise ForbiddenException(detail="Вы не являетесь администратором этой компании")
 
         position = await self.uow.position.get_by_query_one_or_none(name=schema.name)
         if position:
@@ -36,10 +43,17 @@ class PositionService(BaseService):
         return await self.uow.position.add_one_and_get_obj(**data)
 
     @transaction_mode
-    async def update_position(self, position_id: int, schema: PositionUpdateSchema) -> PositionResponseSchema:
+    async def update_position(self,
+                              position_id: int,
+                              schema: PositionUpdateSchema,
+                              user_token_schema: UserTokenSchema) -> PositionResponseSchema:
         position = await self.uow.position.get_by_query_one_or_none(id=position_id)
+
         if not position:
             raise BadRequestException(detail="Такой должности не существует")
+
+        if position.company_id != user_token_schema.company_id:
+            raise ForbiddenException(detail="Вы не являетесь администратором этой компании")
 
         new_position = await self.uow.position.update_one_by_id(obj_id=position_id,
                                                                 name=schema.name,
@@ -47,10 +61,14 @@ class PositionService(BaseService):
         return PositionResponseSchema(id=new_position.id)
 
     @transaction_mode
-    async def delete_position(self, position_id: int) -> DeletePositionSuccessSchema:
+    async def delete_position(self, position_id: int, user_token_schema: UserTokenSchema) -> DeletePositionSuccessSchema:
         position = await self.uow.position.get_by_query_one_or_none(id=position_id)
+
         if not position:
             raise BadRequestException(detail="Такой должности не существует")
+
+        if position.company_id != user_token_schema.company_id:
+            raise ForbiddenException(detail="Вы не являетесь администратором этой компании")
 
         await self.uow.position.delete_by_query(id=position_id)
         return DeletePositionSuccessSchema(message="Должность удалена")
